Video 37. Added permission check for Delete Item button/functionality

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -50,10 +50,18 @@ const Mutations = {
 		const item = await ctx.db.query.item({where}, `{
 			id
 			title
+			user {
+				id
+			}
 		}`);
 
 		// check if they own the item or have the permissions
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
 
+		if (!ownsItem && hasPermissions) {
+			throw new Error('You cannot delete items');
+		}
 
 		// delete item
 		return ctx.db.mutation.deleteItem({where}, info);
